fix crash on linux when loading yml from a file (#38)
* fix crash on linux when loading yml from a file

this issue does not happen on macos, but it does on linux. the issue is
that getFieldName() return a string that has been prefixed with a space
if the indent is non-zero. that causes a crash as when it does the
compare on line 221, the space causes the compare to fail and not match.
added a trimLeadingSpaces()